Working Timezone Change Based on Position Needs Weapon Switching
When the player moves onto an object with a tag with the same name as a zone object (right click Create/Level/Zone) the objects in the scene will change based on the current Zone definitions. If the floor material is set to red on one zone object and blue on another zone object then all floor objects will change to that colour when the player steps on that object. Need to add weapon switching